Pack: Move MbedCrypto configuration file for TF-M Crypto
Preparation for Pack integration:
Moves MbedCrypto configuration file for TF-M Crypto to a standalone folder.
Avoids that the configuration template is referenced from the pack directly.
Configuration file is copied to the project (destination folder is listed
in the include search path, where source folder from the pack is not listed).